Use a thread-safe temporary arena on every platform.
This commit continues the work started in commits 14a571aa and
5358aecd that parallelizes certain geometric operations. This commit
cleans up the temporary arena implementations and makes them
thread-safe.

Also, in commit 14a571aa, a call to FreeAllTemporary() was added
during initialization to create the heap on Windows. This is now
not necessary as the heap is created transparently on the first call
to AllocTemporary().