Revert "CMake: use sanitizer flags for internal targets only"
This reverts commit 5a6994e423471a3e4f6d1a1a552831b2c09d2fcc.

The warnings are valuable and shouldn't be cast aside.
As of e24cf8c, we special case macOS and don't set -fno-sanitize-recover
to allow CI to succeed.
In the future, this could be made stricter again by only suppressing
known bugs, which ideally should also be fixed or reported upstream.