GUI: Allow ESC to finish drawing bezier curves instead of deleting them.
With this pressing ESC while drawing a bezier curve finishes it instead of
deleting it. This reverts the behaviour introduced in 10f2c44ee43bd02cb5ad46c1af8be44aa5c76b99
(p.r. #1386). ESC still deletes the last segment of line segments and
cancels arcs and circles.

This makes it easier to draw non closed curves - otherwise the only way
to finish them was with a right-click.