render: remove a call to `AllocTemporary()` from the OpenGL 1 renderer
This needlessly uses the global temporary arena, when the lifetime can
be managed by replacing `GLU_TESS_COMBINE` with `GLU_TESS_COMBINE_DATA`
instead.

While the use of the global temporary arena is mostly harmless, having
it around needlessly prevents restricting the arena's scope to the sketch
in a possible future where the global sketch is removed. Additionally,
it means that calls to `FreeAllTemporary()` must synchronise with the
renderer, which is a needless restriction on the users of the arena given
the fact that none of the other renderers use it.