slvs: don't try to initially satisfy unsupported constraints
The current code expects a constraint to only generate a single equation
in order to be initially satisfiable. However, some constraints generate
more than a single equation, and this hits an assertion.

SolveSpace itself only initially satisfies a small subset of constraints,
so there is no point in trying to satisfy all types.