emscripten: fix build issues and make the interface functional
Get rid of the bespoke toolchain files, and use the official Emscripten
one instead (it's not needed if an emsdk environment is active).

While at it, fix the build and runtime issues of the Emscripten port so
that it's at least able to run and have a functional interface without
crashing.