build: add a workaround for building zlib in Emscripten
Commit 60076e2fc71233fb5d7ada62d9e16ed7120f98cd updated the Emscripten
toolchain configuration to the official one, and accidentally broke the
build due to a `zlib` issue that was missed.

In the next `zlib` version this can be properly fixed, but until it's
released we can work around it by unsetting `UNIX` for Emscripten right
before including `zlib` and resetting it back again.