build: fix MinGW build failures and update documentation
This includes updating the ANGLE dependency to the latest commit in our
fork which fixes a build issue with GCC.

Additionally, a minor change to the build configuration was needed for
building under 64-bit MinGW, which doesn't support `--large-address-aware`.

With those in place, the documentation for building under MinGW is updated
with working instructions and a recommendation to use MSYS2 for easier
installation of the required tools.