build: fix link error when building on Windows with Clang (not clang-cl)
With CMake >= 3.22, the "-Xlinker /MANIFEST:EMBED" flag is provided even
when no manifest files are added, and since we add our own manifest file
through a resource file, this fails with a duplicate resource error.

Unfortunately there's no easy way to avoid it, other than special casing
for CMake >= 3.22 to add the manifest as a source for non-MINGW WIN32,
and exclude it from the resources, but otherwise add it to the resource
file.

However, this solution requires conditionally reaching into `res/` from
the `src/` directory build, which is really bad, so for now hack out the
offending part from `CMAKE_CXX_LINK_EXECUTABLE`, since we actually don't
want it.