generate: fix `PruneRequests()` to actually prune requests
Commit 86f20cc7e5061d14323fdc0d814980922425e152 (merged in #428) was
marked NFC, even though it didn't just convert loops to ranged-for, but
also introduced a functional change in the request pruning logic to
prune entities instead.

This doesn't make much sense, since the requests will stay, and their
entities will simply be regenerated, leading to an infinite recursion
and resulting in a crash with stack overflow. I'm not sure how it
slipped past code review, but this was the cause of #628 and is the
reason for the additional workplane deletion code that whitequark added
in 61fc345a5d2d4c5773564b2090de6a422e8a39a1, so this commit also reverts
that code as well.