split: fix line splitting for lines intersecting at their endpoints
When two line segments only intersect at their endpoints, the splitting
logic creates four line segments instead of the original two. Two of the
line segments have their endpoints constrained such that they are fully
intersecting and cannot be moved as separate lines, while the original
intersecting endpoint is forced to jump to 0,0,0. This is caused by the
logic not finding an intersection point, but failing to indicate this
case, causing the following code to use the zero vector as the intersection
point.

Since the lines intersect on their endpoints, splitting them doesn't
make much sense, even if the actual intersection point was found,
because it would create zero length line segments. With the created line
segments fully intersecting, and the original point forced to 0, this is
even less useful, so just refuse to split in such a case.